Implemented AssetManager, broke FrameBuffer

+++ SpaghettiEngine/src/main/java/com/spaghettiengine/interfaces/Tickable.java
@@ -2,6 +2,6 @@ package com.spaghettiengine.interfaces;
 
 public interface Tickable {
 
-	public abstract void update(float delta);
+	public abstract void update(double delta);
 
 }
